+++ src/test/java/com/doublew2w/sbs/mybatis/binding/test/IUserDaoApiTest.java
@@ -1,13 +1,15 @@
 package com.doublew2w.sbs.mybatis.binding.test;
 
-import com.doublew2w.sbs.mybatis.binding.MapperProxyFactory;
+import com.doublew2w.sbs.mybatis.binding.MapperRegistry;
+import com.doublew2w.sbs.mybatis.binding.test.dao.IUserDao;
+import com.doublew2w.sbs.mybatis.session.SqlSession;
+import com.doublew2w.sbs.mybatis.session.SqlSessionFactory;
+import com.doublew2w.sbs.mybatis.session.defaults.DefaultSqlSessionFactory;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Proxy;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * @author: DoubleW2w
@@ -19,17 +21,21 @@ public class IUserDaoApiTest {
 
   @Test
   public void test_MapperProxyFactory() {
-    MapperProxyFactory<IUserDao> factory = new MapperProxyFactory<>(IUserDao.class);
-    Map<String, String> sqlSession = new HashMap<>();
-    sqlSession.put(
-        "com.doublew2w.sbs.mybatis.binding.test.IUserDao.queryUserName",
-        "模拟执行 Mapper.xml 中 SQL 语句的操作：查询用户姓名");
-    sqlSession.put(
-        "com.doublew2w.sbs.mybatis.binding.test.IUserDao.queryUserAge",
-        "模拟执行 Mapper.xml 中 SQL 语句的操作：查询用户年龄");
-    IUserDao userDao = factory.newInstance(sqlSession);
+    // 1. 注册 Mapper
+    MapperRegistry registry = new MapperRegistry();
+    // 2. 扫描包路径
+    registry.addMappers("com.doublew2w.sbs.mybatis.binding.test.dao");
+
+    // 创建SqlSession工厂，创建SqlSession
+    SqlSessionFactory sqlSessionFactory = new DefaultSqlSessionFactory(registry);
+    SqlSession sqlSession = sqlSessionFactory.openSession();
+
+    // 3. 获取映射器对象
+    IUserDao userDao = sqlSession.getMapper(IUserDao.class);
+
+    // 4. 测试验证
     String res = userDao.queryUserName("10001");
-    logger.info("测试结果：{}", res);
+    logger.info( res);
   }
 
   @Test