+++ src/test/java/com/doublew2w/sbs/mybatis/binding/test/dao/IUserDao2.java
@@ -1,0 +1,10 @@
+package com.doublew2w.sbs.mybatis.binding.test.dao;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/9/1 6:00
+ * @project: sbs-mybatis
+ */
+public interface IUserDao2 {
+  Integer queryUserAge2(String uId);
+}